requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.5
Django==4.0.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
filelock==3.7.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.2
platformdirs==2.5.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.2
tqdm==4.65.0
tzdata==2022.1
urllib3==1.26.15
virtualenv==20.16.3
virtualenvwrapper-win==1.2.7
yarl==1.8.2
